infra/lambda/image_processor/requirements.txt
Pillow==10.1.0
boto3==1.28.0
